Fix two display issues. Autocomputed and existing sizes were truncated to too few digits, which sometimes led to rounding errors when adding new partitions. Also, on systems with 32-bit size_t, the main panel could truncate the medium size, producing nonsense for the displayed sizes of large devices.

+++ partedit/diskeditor.h
@@ -1,7 +1,9 @@
+#include <inttypes.h>
+
 struct partedit_item {
 	int indentation;
 	const char *name;
-	size_t size;
+	intmax_t size;
 	const char *type;
 	char *mountpoint;
